SDL3 Migration
Migrating from SDL2 to SDL3, now requiring SDL3 libraries instead.

+++ include/cosmac_vip.hpp
@@ -181,7 +181,7 @@ namespace VIPR_Emulator
 
 			inline void AdjustRAM(uint8_t RAM_KB)
 			{
-				RAM.resize(RAM_KB << 10);
+				RAM.resize(static_cast<size_t>(RAM_KB) << 10);
 				memset(RAM.data(), 0, RAM.size());
 			}
